BUG FIX: Fixed a bug in 'file_io' where size for the file reading was a long so it could go negative so we have prevented that using a 'size_t' instead.

+++ src/file_io.h
@@ -1,9 +1,10 @@
 #ifndef ALY_COMPILER_FILE_IO_H
 #define ALY_COMPILER_FILE_IO_H
 
+#include <stddef.h>
 #include <stdio.h>
 
-long file_size(FILE* file);
+size_t file_size(FILE* file);
 char* file_contents(char* path);
 
 #endif // ALY_COMPILER_FILE_IO_H
